Промежуточные правки
